introduce geometry subpackage
