+++ artist/util/utils.py
@@ -9,517 +9,6 @@ from artist.util import config_dictionary, index_mapping
 from artist.util.environment_setup import get_device
 
 
-def rotate_distortions(
-    e: torch.Tensor,
-    u: torch.Tensor,
-    device: torch.device | None = None,
-) -> torch.Tensor:
-    """
-    Rotate the distortions for the light source.
-
-    Rotate around the up and then the east axis in this very order in a right-handed east-north-up
-    coordinate system. Positive angles result in a rotation in the mathematical direction of rotation, i.e.,
-    counter-clockwise. Points need to be multiplied as column vectors from the right-hand side with the
-    resulting rotation matrix. Note that the order is fixed due to the non-commutative property of matrix-matrix
-    multiplication.
-
-    Parameters
-    ----------
-    e : torch.Tensor
-        East rotation angles in radians.
-        Shape is ``[number_of_heliostats, number_of_rays, number_of_surface_points]``.
-    u : torch.Tensor
-        Up rotation angles in radians.
-        Shape is ``[number_of_heliostats, number_of_rays, number_of_surface_points]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Raises
-    ------
-    ValueError
-        If the shapes of the input tensors do not match.
-
-    Returns
-    -------
-    torch.Tensor
-        Batched 4×4 rotation matrices, one per distortion sample.
-        Shape is ``[number_of_heliostats, number_of_rays, number_of_surface_points, 4, 4]``.
-    """
-    device = get_device(device=device)
-
-    if e.shape != u.shape:
-        raise ValueError(
-            "The two tensors containing angles for the east and up rotation must have the same shape."
-        )
-
-    cos_e = torch.cos(e)
-    sin_e = -torch.sin(e)  # Heliostat convention.
-    cos_u = torch.cos(u)
-    sin_u = torch.sin(u)
-    ones = torch.ones_like(e, device=device)
-
-    matrix = torch.zeros(
-        e.shape[index_mapping.heliostat_dimension],
-        e.shape[index_mapping.facet_dimension],
-        e.shape[index_mapping.points_dimension],
-        4,
-        4,
-        device=device,
-    )
-
-    matrix[:, :, :, index_mapping.e, index_mapping.e] = cos_u
-    matrix[:, :, :, index_mapping.e, index_mapping.n] = -sin_u
-    matrix[:, :, :, index_mapping.n, index_mapping.e] = cos_e * sin_u
-    matrix[:, :, :, index_mapping.n, index_mapping.n] = cos_e * cos_u
-    matrix[:, :, :, index_mapping.n, index_mapping.u] = sin_e
-    matrix[:, :, :, index_mapping.u, index_mapping.e] = -sin_e * sin_u
-    matrix[:, :, :, index_mapping.u, index_mapping.n] = -sin_e * cos_u
-    matrix[:, :, :, index_mapping.u, index_mapping.u] = cos_e
-    matrix[
-        :,
-        :,
-        :,
-        index_mapping.transform_homogeneous,
-        index_mapping.transform_homogeneous,
-    ] = ones
-
-    return matrix
-
-
-def rotate_e(
-    e: torch.Tensor,
-    device: torch.device | None = None,
-) -> torch.Tensor:
-    """
-    Rotate around the east axis.
-
-    Rotate around the east axis in a right-handed east-north-up coordinate system. Positive angles result in a rotation
-    in the mathematical direction of rotation, i.e., counter-clockwise. Points need to be multiplied as column vectors
-    from the right-hand side with the resulting rotation matrix.
-
-    Parameters
-    ----------
-    e : torch.Tensor
-        East rotation angles in radians.
-        Shape is ``[number_of_heliostats]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        Batched 4×4 east-axis rotation matrices, one per heliostat.
-        Shape is ``[number_of_heliostats, 4, 4]``.
-    """
-    device = get_device(device=device)
-
-    cos_e = torch.cos(e)
-    sin_e = torch.sin(e)
-    ones = torch.ones_like(e, device=device)
-
-    matrix = torch.zeros(
-        e.shape[index_mapping.heliostat_dimension], 4, 4, device=device
-    )
-
-    matrix[:, index_mapping.e, index_mapping.e] = ones
-    matrix[:, index_mapping.n, index_mapping.n] = cos_e
-    matrix[:, index_mapping.n, index_mapping.u] = -sin_e
-    matrix[:, index_mapping.u, index_mapping.n] = sin_e
-    matrix[:, index_mapping.u, index_mapping.u] = cos_e
-    matrix[
-        :, index_mapping.transform_homogeneous, index_mapping.transform_homogeneous
-    ] = ones
-
-    return matrix
-
-
-def rotate_n(n: torch.Tensor, device: torch.device | None = None) -> torch.Tensor:
-    """
-    Rotate around the north axis.
-
-    Rotate around the north axis in a right-handed east-north-up coordinate system. Positive angles result in a rotation
-    in the mathematical direction of rotation, i.e., counter-clockwise. Points need to be multiplied as column vectors
-    from the right-hand side with the resulting rotation matrix.
-
-    Parameters
-    ----------
-    n : torch.Tensor
-        North rotation angles in radians.
-        Shape is ``[number_of_heliostats]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        Batched 4×4 north-axis rotation matrices, one per heliostat.
-        Shape is ``[number_of_heliostats, 4, 4]``.
-    """
-    device = get_device(device=device)
-
-    cos_n = torch.cos(n)
-    sin_n = torch.sin(n)
-    ones = torch.ones_like(n, device=device)
-
-    matrix = torch.zeros(
-        n.shape[index_mapping.heliostat_dimension], 4, 4, device=device
-    )
-
-    matrix[:, index_mapping.e, index_mapping.e] = cos_n
-    matrix[:, index_mapping.e, index_mapping.u] = -sin_n
-    matrix[:, index_mapping.n, index_mapping.n] = ones
-    matrix[:, index_mapping.u, index_mapping.e] = sin_n
-    matrix[:, index_mapping.u, index_mapping.u] = cos_n
-    matrix[
-        :, index_mapping.transform_homogeneous, index_mapping.transform_homogeneous
-    ] = ones
-
-    return matrix
-
-
-def rotate_u(u: torch.Tensor, device: torch.device | None = None) -> torch.Tensor:
-    """
-    Rotate around the up axis.
-
-    Rotate around the up axis in a right-handed east-north-up coordinate system. Positive angles result in a rotation in
-    the mathematical direction of rotation, i.e., counter-clockwise. Points need to be multiplied as column vectors from
-    the right-hand side with the resulting rotation matrix.
-
-    Parameters
-    ----------
-    u : torch.Tensor
-        Up rotation angles in radians.
-        Shape is ``[number_of_heliostats]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        Batched 4×4 up-axis rotation matrices, one per heliostat.
-        Shape is ``[number_of_heliostats, 4, 4]``.
-    """
-    device = get_device(device=device)
-
-    cos_u = torch.cos(u)
-    sin_u = torch.sin(u)
-    ones = torch.ones_like(u, device=device)
-
-    matrix = torch.zeros(
-        u.shape[index_mapping.heliostat_dimension], 4, 4, device=device
-    )
-
-    matrix[:, index_mapping.e, index_mapping.e] = cos_u
-    matrix[:, index_mapping.e, index_mapping.n] = -sin_u
-    matrix[:, index_mapping.n, index_mapping.e] = sin_u
-    matrix[:, index_mapping.n, index_mapping.n] = cos_u
-    matrix[:, index_mapping.u, index_mapping.u] = ones
-    matrix[
-        :, index_mapping.transform_homogeneous, index_mapping.transform_homogeneous
-    ] = ones
-
-    return matrix
-
-
-def translate_enu(
-    e: torch.Tensor,
-    n: torch.Tensor,
-    u: torch.Tensor,
-    device: torch.device | None = None,
-) -> torch.Tensor:
-    """
-    Translate in all directions.
-
-    Translate a given point in the east, north, and up direction. Note that the point must be multiplied as a column
-    vector from the right-hand side of the resulting matrix.
-
-    Parameters
-    ----------
-    e : torch.Tensor
-        East translation distances in meters.
-        Shape is ``[number_of_heliostats]``.
-    n : torch.Tensor
-        North translation distances in meters.
-        Shape is ``[number_of_heliostats]``.
-    u : torch.Tensor
-        Up translation distances in meters.
-        Shape is ``[number_of_heliostats]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Raises
-    ------
-    ValueError
-        If the sizes of the input tensors do not match.
-
-    Returns
-    -------
-    torch.Tensor
-        Batched 4×4 translation matrices, one per heliostat.
-        Shape is ``[number_of_heliostats, 4, 4]``.
-    """
-    device = get_device(device=device)
-
-    if not (e.shape == u.shape == n.shape):
-        raise ValueError(
-            "The three tensors containing the east, north, and up translations must have the same shape."
-        )
-
-    ones = torch.ones_like(e, device=device)
-
-    matrix = torch.zeros(
-        e.shape[index_mapping.heliostat_dimension], 4, 4, device=device
-    )
-
-    matrix[:, index_mapping.e, index_mapping.e] = ones
-    matrix[:, index_mapping.e, index_mapping.transform_homogeneous] = e
-    matrix[:, index_mapping.n, index_mapping.n] = ones
-    matrix[:, index_mapping.n, index_mapping.transform_homogeneous] = n
-    matrix[:, index_mapping.u, index_mapping.u] = ones
-    matrix[:, index_mapping.u, index_mapping.transform_homogeneous] = u
-    matrix[
-        :, index_mapping.transform_homogeneous, index_mapping.transform_homogeneous
-    ] = ones
-
-    return matrix
-
-
-def convert_3d_points_to_4d_format(
-    points: torch.Tensor, device: torch.device | None = None
-) -> torch.Tensor:
-    """
-    Append ones to the last dimension of 3D point vectors.
-
-    Includes the convention that points have a 1 and directions have a 0 as 4th dimension.
-    This function can handle batched tensors.
-
-    Parameters
-    ----------
-    points : torch.Tensor
-        Input points in a 3D format.
-        Shape is ``[..., 3]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Raises
-    ------
-    ValueError
-        If the input is not 3D.
-
-    Returns
-    -------
-    torch.Tensor
-        Point vector with ones appended at the last dimension.
-        Shape is ``[..., 4]``.
-    """
-    device = get_device(device=device)
-
-    if points.size(dim=-1) != 3:
-        raise ValueError(f"Expected 3D points but got points of shape {points.shape}!")
-
-    ones_tensor = torch.ones(
-        points.shape[:-1] + (1,), dtype=points.dtype, device=device
-    )
-    return torch.cat((points, ones_tensor), dim=-1)
-
-
-def convert_3d_directions_to_4d_format(
-    directions: torch.Tensor, device: torch.device | None = None
-) -> torch.Tensor:
-    """
-    Append zeros to the last dimension of 3D direction vectors.
-
-    Includes the convention that points have a 1 and directions have a 0 as 4th dimension.
-    This function can handle batched tensors.
-
-    Parameters
-    ----------
-    directions : torch.Tensor
-        Input direction in a 3D format.
-        Shape is ``[..., 3]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Raises
-    ------
-    ValueError
-        If the input is not 3D.
-
-    Returns
-    -------
-    torch.Tensor
-        Direction vectors with zeros appended at the last dimension.
-        Shape is ``[..., 4]``.
-    """
-    device = get_device(device=device)
-
-    if directions.size(dim=-1) != 3:
-        raise ValueError(
-            f"Expected 3D directions but got directions of shape {directions.shape}!"
-        )
-
-    zeros_tensor = torch.zeros(
-        directions.shape[:-1] + (1,), dtype=directions.dtype, device=device
-    )
-    return torch.cat((directions, zeros_tensor), dim=-1)
-
-
-def normalize_points(points: torch.Tensor) -> torch.Tensor:
-    """
-    Normalize each column of a 2D tensor to the open interval (0, 1).
-
-    Parameters
-    ----------
-    points : torch.Tensor
-        A tensor containing points to be normalized.
-        Shape is ``[number_of_points, number_of_dimensions]``.
-
-    Returns
-    -------
-    torch.Tensor
-        The normalized points.
-        Shape is ``[number_of_points, number_of_dimensions]``.
-    """
-    # Since we want the open interval (0,1), a small offset is required to also exclude the boundaries.
-    min_vals = torch.min(points, dim=index_mapping.unbatched_tensor_values).values
-    point_range = points - min_vals
-    max_vals = torch.max(
-        point_range + 2e-5, dim=index_mapping.unbatched_tensor_values
-    ).values
-    return (point_range + 1e-5) / max_vals
-
-
-def decompose_rotations(
-    initial_vector: torch.Tensor,
-    target_vector: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """
-    Compute ENU components of the axis-angle rotation vector that rotates initial vectors toward a target vector.
-
-    This function does **not** perform an Euler-angle decomposition. Instead, it computes:
-    1) the rotation axis via cross product, and
-    2) the rotation magnitude via arccos of the dot product,
-    then returns the Cartesian components of the rotation vector
-    (`theta * axis`) in east, north, and up coordinates.
-
-    Parameters
-    ----------
-    initial_vector : torch.Tensor
-        Initial vectors in homogeneous coordinates.
-        Shape is ``[number_of_heliostats, 4]``.
-        Only the first three components (ENU) are used.
-    target_vector : torch.Tensor
-        Target vector in homogeneous coordinates.
-        Shape is ``[4]``.
-        Only the first three components (ENU) are used.
-
-    Returns
-    -------
-    torch.Tensor
-        East component of the axis-angle rotation vector.
-        Shape is ``[number_of_heliostats]``.
-    torch.Tensor
-        North component of the axis-angle rotation vector.
-        Shape is ``[number_of_heliostats]``.
-    torch.Tensor
-        Up component of the axis-angle rotation vector.
-        Shape is ``[number_of_heliostats]``.
-    """
-    # Normalize the input vectors.
-    initial_vector = torch.nn.functional.normalize(
-        initial_vector[:, : index_mapping.slice_fourth_dimension]
-    )
-    target_vector = torch.nn.functional.normalize(
-        target_vector[: index_mapping.slice_fourth_dimension],
-        dim=index_mapping.unbatched_tensor_values,
-    ).unsqueeze(index_mapping.unbatched_tensor_values)
-
-    # Compute the cross product (rotation axis).
-    r = torch.linalg.cross(initial_vector, target_vector)
-
-    # Normalize the rotation axis.
-    r_normalized = torch.nn.functional.normalize(r)
-
-    # Compute the angle between the vectors.
-    theta = torch.arccos(torch.clamp(initial_vector @ target_vector.T, -1.0, 1.0))
-
-    # Decompose the angle along each axis.
-    theta_components = theta * r_normalized
-
-    return theta_components[:, 0], theta_components[:, 1], theta_components[:, 2]
-
-
-def rotation_angle_and_axis(
-    from_orientation: torch.Tensor,
-    to_orientation: torch.Tensor,
-    device: torch.device | None = None,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """
-    Compute the rotation axis and angle between two orientations.
-
-    Parameters
-    ----------
-    from_orientation : torch.Tensor
-        The original orientation.
-        Shape is ``[4]``.
-    to_orientation : torch.Tensor
-        The rotated orientation.
-        Shape is ``[4]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        The rotation axis.
-        Shape is ``[3]``.
-    torch.Tensor
-        The angle of the rotation as a scalar tensor.
-    """
-    device = get_device(device=device)
-    from_orientation = from_orientation[:3] / torch.norm(from_orientation[:3])
-    to_orientation = to_orientation[:3] / torch.norm(to_orientation[:3])
-    dot = torch.clamp(torch.dot(from_orientation, to_orientation), -1.0, 1.0)
-    angle = torch.acos(dot)
-    axis = torch.linalg.cross(from_orientation, to_orientation)
-    axis_norm = torch.norm(axis)
-    # Parallel vectors.
-    epsilon = 1e-6
-    if axis_norm < epsilon and dot > 0:
-        return torch.tensor([1.0, 0.0, 0.0], device=device), torch.tensor(
-            0.0, device=device
-        )
-    # Inverse vectors.
-    if axis_norm < epsilon and dot < 0:
-        if abs(from_orientation[index_mapping.e]) < abs(
-            from_orientation[index_mapping.n]
-        ):
-            orthogonal = torch.tensor([1.0, 0.0, 0.0], device=device)
-        else:
-            orthogonal = torch.tensor([0.0, 1.0, 0.0], device=device)
-        axis = torch.linalg.cross(from_orientation, orthogonal)
-        axis = axis / torch.norm(axis)
-        return axis, torch.tensor(torch.pi, device=device)
-    axis = axis / axis_norm
-    return axis, angle
-
-
 def get_center_of_mass(
     bitmaps: torch.Tensor,
     device: torch.device | None = None,
@@ -580,220 +69,6 @@ def get_center_of_mass(
     )
 
     return torch.stack([e_center_of_mass, u_center_of_mass], dim=1)
-
-
-def bitmap_coordinates_to_target_coordinates(
-    bitmap_coordinates: torch.Tensor,
-    bitmap_resolution: torch.Tensor,
-    solar_tower: SolarTower,
-    target_area_indices: torch.Tensor,
-    device: torch.device | None = None,
-) -> torch.Tensor:
-    """
-    Convert bitmap pixel coordinates to 4D homogeneous world coordinates on the target surface.
-
-    For planar target areas the pixel coordinates are mapped linearly to target plane coordinates.
-    For cylindrical target areas the pixel coordinates are mapped to cylindrical surface coordinates
-    using the cylinder's radius, opening angle, height, axis, and normal.
-
-    Bitmaps and the resolution are conceptually defined as: [W, H] # width, height
-    Tensor memory layout follows PyTorch convention: [H, W] # height, width
-
-    The bitmap is treated as a discrete image grid with resolution:
-        - bitmap_resolution = [width, height]
-    Pixel coordinates follow image indexing conventions:
-        - bitmap_coordinates[..., e] ∈ [0, W-1]
-        - bitmap_coordinates[..., u] ∈ [0, H-1]
-    They are interpreted as centered pixels:
-        - (e + 0.5) / W
-        - (u + 0.5) / H
-    This ensures each pixel represents its spatial cell center rather than its corner.
-
-    The e-axis is intentionally flipped (0.5 - e_norm) to match the desired bitmap orientation.
-    This means: increasing bitmap e → decreases world e.
-
-    Parameters
-    ----------
-    bitmap_coordinates : torch.Tensor
-        Pixel coordinates in the bitmap for each heliostat, as (e, u) pairs.
-        Shape is ``[number_of_active_heliostats, 2]``.
-    bitmap_resolution : torch.Tensor
-        Resolution of the bitmap (width, height) in pixels.
-        Shape is ``[2]``.
-    solar_tower : SolarTower
-        Solar tower containing all target area definitions (planar and cylindrical).
-    target_area_indices : torch.Tensor
-        Global target area index for each heliostat (planar indices first, cylindrical second).
-        Shape is ``[number_of_active_heliostats]``.
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        World coordinates on the target surface in homogeneous format.
-        Shape is ``[number_of_active_heliostats, 4]``.
-    """
-    device = get_device(device=device)
-
-    number_of_coordinates = target_area_indices.numel()
-    target_coordinates = torch.zeros((number_of_coordinates, 4), device=device)
-    target_coordinates[:, -1] = 1.0
-
-    bitmap_width = bitmap_resolution[index_mapping.unbatched_bitmap_e]
-    bitmap_height = bitmap_resolution[index_mapping.unbatched_bitmap_u]
-
-    e_norm = (
-        bitmap_coordinates[:, index_mapping.unbatched_bitmap_e] + 0.5
-    ) / bitmap_width
-    u_norm = (
-        bitmap_coordinates[:, index_mapping.unbatched_bitmap_u] + 0.5
-    ) / bitmap_height
-
-    planar_mask = (
-        target_area_indices
-        < solar_tower.number_of_target_areas_per_type[index_mapping.planar_target_areas]
-    )
-
-    if planar_mask.any():
-        planar_indices = target_area_indices[planar_mask]
-
-        planar = cast(
-            TowerTargetAreasPlanar,
-            solar_tower.target_areas[index_mapping.planar_target_areas],
-        )
-
-        centers = planar.centers[planar_indices][:, :3]
-        dims = planar.dimensions[planar_indices]
-
-        e_axis = torch.tensor([1.0, 0.0, 0.0], device=device).expand(
-            planar_indices.numel(), 3
-        )
-        u_axis = torch.tensor([0.0, 0.0, 1.0], device=device).expand(
-            planar_indices.numel(), 3
-        )
-
-        e_local = (0.5 - e_norm[planar_mask]) * dims[
-            :, index_mapping.target_dimensions_width
-        ]
-        u_local = (0.5 - u_norm[planar_mask]) * dims[
-            :, index_mapping.target_dimensions_height
-        ]
-
-        target_coordinates[planar_mask, :3] = (
-            centers + e_local[:, None] * e_axis + u_local[:, None] * u_axis
-        )
-
-    if (~planar_mask).any():
-        cylinder_indices = (
-            target_area_indices[~planar_mask]
-            - solar_tower.number_of_target_areas_per_type[
-                index_mapping.planar_target_areas
-            ]
-        )
-        cylindrical = cast(
-            TowerTargetAreasCylindrical,
-            solar_tower.target_areas[index_mapping.cylindrical_target_areas],
-        )
-
-        centers = cylindrical.centers[cylinder_indices][:, :3]
-        axes = cylindrical.axes[cylinder_indices][:, :3]
-        normals = cylindrical.normals[cylinder_indices][:, :3]
-        radii = cylindrical.radii[cylinder_indices].flatten()
-        heights = cylindrical.heights[cylinder_indices].flatten()
-        opening_angles = cylindrical.opening_angles[cylinder_indices].flatten()
-
-        v = torch.cross(axes, normals, dim=-1)
-
-        theta = (e_norm[~planar_mask] - 0.5) * opening_angles
-        z = (0.5 - u_norm[~planar_mask]) * heights
-
-        target_coordinates[~planar_mask, :3] = (
-            centers
-            + radii[:, None] * torch.cos(theta)[:, None] * normals
-            + radii[:, None] * torch.sin(theta)[:, None] * v
-            + z[:, None] * axes
-        )
-
-    return target_coordinates
-
-
-def perform_canting(
-    canting_angles: torch.Tensor,
-    data: torch.Tensor,
-    inverse: bool = False,
-    device: torch.device | None = None,
-) -> torch.Tensor:
-    """
-    Perform canting (rotation) on data like surface points or surface normals.
-
-    Parameters
-    ----------
-    canting_angles : torch.Tensor
-        Canting angles.
-        Shape is ``[number_of_surfaces, number_of_facets, 2, 4]``.
-    data : torch.Tensor
-        Data to be canted.
-        Shape is ``[number_of_surfaces, number_of_facets, number_of_points_per_facet, 4]``.
-    inverse : bool
-        Indicates the direction of the rotation.
-        Use ``inverse=False`` for canting and ``inverse=True`` for decanting (default is False).
-    device : torch.device | None
-        The device on which to perform computations or load tensors and models (default is None).
-        If None, ``ARTIST`` will automatically select the most appropriate
-        device (CUDA or CPU) based on availability and OS.
-
-    Returns
-    -------
-    torch.Tensor
-        The (de-)canted data.
-        Shape is ``[number_of_surfaces, number_of_facets, number_of_points_per_facet, 4]``.
-    """
-    device = get_device(device=device)
-    canting_angles = canting_angles.to(device)
-    data = data.to(device)
-
-    number_of_surfaces = data.shape[index_mapping.heliostat_dimension]
-    number_of_facets_per_surface = data.shape[index_mapping.facet_dimension]
-
-    rotation_matrix = torch.zeros(
-        (number_of_surfaces, number_of_facets_per_surface, 4, 4),
-        device=device,
-        dtype=data.dtype,
-    )
-
-    # Extract ENU basis candidates from canting tensor (drop homogeneous component).
-    e = canting_angles[:, :, index_mapping.e, : index_mapping.slice_fourth_dimension]
-    n = canting_angles[:, :, index_mapping.n, : index_mapping.slice_fourth_dimension]
-
-    # Build a numerically stable orthonormal basis:
-    # 1) normalize e
-    e = torch.nn.functional.normalize(e, dim=-1)
-    # 2) u = normalize(e x n)
-    u = torch.linalg.cross(e, n, dim=-1)
-    u = torch.nn.functional.normalize(u, dim=-1, eps=1e-8)
-    # 3) n_ortho = normalize(u x e)
-    n_ortho = torch.linalg.cross(u, e, dim=-1)
-    n_ortho = torch.nn.functional.normalize(n_ortho, dim=-1, eps=1e-8)
-
-    # Fill rotation matrix columns with ENU basis vectors.
-    rotation_matrix[:, :, : index_mapping.slice_fourth_dimension, index_mapping.e] = e
-    rotation_matrix[:, :, : index_mapping.slice_fourth_dimension, index_mapping.n] = (
-        n_ortho
-    )
-    rotation_matrix[:, :, : index_mapping.slice_fourth_dimension, index_mapping.u] = u
-    rotation_matrix[
-        :, :, index_mapping.transform_homogeneous, index_mapping.transform_homogeneous
-    ] = 1.0
-
-    # Data is represented as row vectors (..., 4); therefore:
-    # - forward canting uses R^T
-    # - inverse canting uses R
-    if inverse:
-        return data @ rotation_matrix
-    return data @ rotation_matrix.mT
 
 
 def trapezoid_distribution(
