+++ tests/geometry/test_transforms.py
@@ -1,0 +1,412 @@
+import math
+
+import pytest
+import torch
+
+from artist.geometry import transforms
+
+
+@pytest.mark.parametrize(
+    "east_translation, north_translation, up_translation, expected",
+    [
+        (
+            torch.tensor([1.0]),
+            torch.tensor([2.0]),
+            torch.tensor([3.0]),
+            torch.tensor(
+                [
+                    [
+                        [1.0, 0.0, 0.0, 1.0],
+                        [0.0, 1.0, 0.0, 2.0],
+                        [0.0, 0.0, 1.0, 3.0],
+                        [0.0, 0.0, 0.0, 1.0],
+                    ]
+                ]
+            ),
+        ),
+        (
+            torch.tensor([1.0, 0.0, 0.0]),
+            torch.tensor([0.0, 1.0, 0.0, 1.0]),
+            torch.tensor([0.0, 0.0, 1.0, 1.0]),
+            None,
+        ),
+        (
+            torch.tensor([1.0, 0.0, 0.0, 0.0]),
+            torch.tensor([0.0, 1.0, 0.0]),
+            torch.tensor([0.0, 0.0, 1.0, 1.0]),
+            None,
+        ),
+        (
+            torch.tensor([1.0, 0.0, 0.0, 0.0]),
+            torch.tensor([0.0, 1.0, 0.0, 0.0]),
+            torch.tensor([0.0, 0.0, 1.0]),
+            None,
+        ),
+    ],
+)
+def test_translate_enu(
+    east_translation: torch.Tensor,
+    north_translation: torch.Tensor,
+    up_translation: torch.Tensor,
+    expected: torch.Tensor,
+    device: torch.device,
+) -> None:
+    """
+    Test that the correct translation matrix is created.
+
+    Parameters
+    ----------
+    east_translation : torch.Tensor
+        The translation in east direction.
+    north_translation : torch.Tensor
+        The translation in north direction.
+    up_translation : torch.Tensor
+        The translation in up direction.
+    expected : torch.Tensor
+        The expected overall translation or ``None`` if an error is expected.
+    device : torch.device
+        The device on which to initialize tensors.
+
+    Raises
+    ------
+    AssertionError
+        If test does not complete as expected.
+    """
+    # Check if the ValueError is thrown as expected.
+    if expected is None:
+        with pytest.raises(ValueError) as exc_info:
+            transforms.translate_enu(
+                e=east_translation.to(device),
+                n=north_translation.to(device),
+                u=up_translation.to(device),
+                device=device,
+            )
+        assert (
+            "The three tensors containing the east, north, and up translations must have the same shape."
+            in str(exc_info.value)
+        )
+    else:
+        # Check if the translation matrix is correct.
+        translation_matrix = transforms.translate_enu(
+            e=east_translation.to(device),
+            n=north_translation.to(device),
+            u=up_translation.to(device),
+            device=device,
+        )
+        torch.testing.assert_close(
+            translation_matrix, expected.to(device), rtol=1e-4, atol=1e-4
+        )
+
+
+@pytest.mark.parametrize(
+    "e_distortions, u_distortions, rays_to_rotate, expected_distorted_rays",
+    [
+        (  # Rotate single ray in the east direction.
+            torch.tensor([[[math.pi / 2]]]),
+            torch.tensor([[[0.0]]]),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor([[[[1.0, -1.0, 1.0, 0.0]]]]),
+        ),
+        (  # Rotate single ray in the up direction.
+            torch.tensor([[[0.0]]]),
+            torch.tensor([[[math.pi / 2]]]),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor([[[[-1.0, 1.0, 1.0, 0.0]]]]),
+        ),
+        (  # Rotate single ray in the east and then up direction.
+            torch.tensor([[[math.pi / 2]]]),
+            torch.tensor([[[math.pi / 2]]]),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor([[[[-1.0, -1.0, 1.0, 0.0]]]]),
+        ),
+        (  # Consider multiple rotations for a single ray in the east direction.
+            torch.tensor(
+                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
+            ),
+            torch.tensor([[[0]], [[0]], [[0]], [[0]]]),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor(
+                [
+                    [[[1.000000, -1.000000, 1.000000, 0.000000]]],
+                    [[[1.000000, -1.000000, -1.000000, 0.000000]]],
+                    [[[1.000000, 0.000000, 1.414214, 0.000000]]],
+                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for a single ray in the up direction.
+            torch.tensor([[[0]], [[0]], [[0]], [[0]]]),
+            torch.tensor(
+                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
+            ),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor(
+                [
+                    [[[-1.000000, 1.000000, 1.000000, 0.000000]]],
+                    [[[-1.000000, -1.000000, 1.000000, 0.000000]]],
+                    [[[0.000000, 1.414214, 1.000000, 0.000000]]],
+                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for a single ray in the east and then up direction.
+            torch.tensor(
+                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
+            ),
+            torch.tensor(
+                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
+            ),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            torch.tensor(
+                [
+                    [[[-1.000000, -1.000000, 1.000000, 0.000000]]],
+                    [[[-1.000000, 1.000000, -1.000000, 0.000000]]],
+                    [[[0.000000, 0.292893, 1.707107, 0.000000]]],
+                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for multiple rays in the east direction.
+            torch.tensor(
+                [
+                    [[math.pi, math.pi / 2, math.pi / 4]],
+                    [[math.pi / 4, math.pi, math.pi / 2]],
+                ]
+            ),
+            torch.tensor([[[0, 0, 0]], [[0, 0, 0]]]),
+            torch.tensor(
+                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
+            ),
+            torch.tensor(
+                [
+                    [
+                        [
+                            [1.000000e00, -9.999999e-01, -1.000000e00, 0.000000e00],
+                            [2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [3.000000e00, -5.960464e-08, 4.242640e00, 0.000000e00],
+                        ]
+                    ],
+                    [
+                        [
+                            [1.000000e00, 0.000000e00, 1.414214e00, 0.000000e00],
+                            [2.000000e00, -2.000000e00, -2.000000e00, 0.000000e00],
+                            [3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ]
+                    ],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for multiple rays in the up direction.
+            torch.tensor([[[0, 0, 0]], [[0, 0, 0]]]),
+            torch.tensor(
+                [
+                    [[math.pi, math.pi / 2, math.pi / 4]],
+                    [[math.pi / 4, math.pi, math.pi / 2]],
+                ]
+            ),
+            torch.tensor(
+                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
+            ),
+            torch.tensor(
+                [
+                    [
+                        [
+                            [-9.999999e-01, -1.000000e00, 1.000000e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 4.242640e00, 3.000000e00, 0.000000e00],
+                        ]
+                    ],
+                    [
+                        [
+                            [0.000000e00, 1.414214e00, 1.000000e00, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-3.000000e00, 3.000000e00, 3.000000e00, 0.000000e00],
+                        ]
+                    ],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for multiple rays in the east and then up direction.
+            torch.tensor(
+                [
+                    [[math.pi, math.pi / 2, math.pi / 4]],
+                    [[math.pi / 4, math.pi, math.pi / 2]],
+                ]
+            ),
+            torch.tensor(
+                [
+                    [[math.pi, math.pi / 2, math.pi / 4]],
+                    [[math.pi / 4, math.pi, math.pi / 2]],
+                ]
+            ),
+            torch.tensor(
+                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
+            ),
+            torch.tensor(
+                [
+                    [
+                        [
+                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
+                        ]
+                    ],
+                    [
+                        [
+                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
+                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ]
+                    ],
+                ]
+            ),
+        ),
+        (  # Consider multiple rotations for multiple rays on four facets in the east and then up direction.
+            torch.tensor(
+                [
+                    [
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                    ],
+                    [
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                    ],
+                ]
+            ),
+            torch.tensor(
+                [
+                    [
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                        [math.pi, math.pi / 2, math.pi / 4],
+                    ],
+                    [
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                        [math.pi / 4, math.pi, math.pi / 2],
+                    ],
+                ]
+            ),
+            torch.tensor(
+                [
+                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
+                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
+                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
+                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
+                ]
+            ),
+            torch.tensor(
+                [
+                    [
+                        [
+                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
+                        ],
+                        [
+                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
+                        ],
+                        [
+                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
+                        ],
+                        [
+                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
+                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
+                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
+                        ],
+                    ],
+                    [
+                        [
+                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
+                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ],
+                        [
+                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
+                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ],
+                        [
+                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
+                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ],
+                        [
+                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
+                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
+                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
+                        ],
+                    ],
+                ]
+            ),
+        ),
+        (  # Test raise ValueError
+            torch.tensor([[[math.pi / 2]]]),
+            torch.tensor([[[0.0], [0.0]]]),
+            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
+            None,
+        ),
+    ],
+)
+def test_distortion_rotations(
+    e_distortions: torch.Tensor,
+    u_distortions: torch.Tensor,
+    rays_to_rotate: torch.Tensor,
+    expected_distorted_rays: torch.Tensor,
+    device: torch.device,
+) -> None:
+    """
+    Test the rotation function used for scattering rays by considering various rotations.
+
+    Parameters
+    ----------
+    e_distortions : torch.Tensor
+        The distortions in the east direction used in the rotation matrix.
+    u_distortions : torch.Tensor
+        The distortions in the upper direction used in the rotation matrix.
+    rays_to_rotate : torch.Tensor
+        The rays to rotate given the distortions.
+    expected_distorted_rays : torch.Tensor
+        The expected distorted rays after rotation.
+    device : torch.device
+        The device on which to initialize tensors.
+
+    Raises
+    ------
+    AssertionError
+        If test does not complete as expected.
+    """
+    if expected_distorted_rays is None:
+        with pytest.raises(ValueError) as exc_info:
+            distorted_rays = (
+                transforms.rotate_distortions(
+                    e=e_distortions.to(device),
+                    u=u_distortions.to(device),
+                    device=device,
+                )
+                @ rays_to_rotate.to(device).unsqueeze(-1)
+            ).squeeze(-1)
+        assert (
+            "The two tensors containing angles for the east and up rotation must have the same shape."
+            in str(exc_info.value)
+        )
+    else:
+        distorted_rays = (
+            transforms.rotate_distortions(
+                e=e_distortions.to(device), u=u_distortions.to(device), device=device
+            )
+            @ rays_to_rotate.to(device).unsqueeze(-1)
+        ).squeeze(-1)
+
+        torch.testing.assert_close(distorted_rays, expected_distorted_rays.to(device))
