+++ tests/nurbs/test_surfaces.py
@@ -2,6 +2,7 @@ import torch
 
 import artist.nurbs.utils
 import artist.util.utils
+from artist.geometry import coordinates, transforms
 from artist.nurbs.surfaces import NURBSSurfaces
 
 
@@ -90,7 +91,7 @@ def test_nurbs(device: torch.device) -> None:
     )
     facet_translation = torch.tensor([[[[0.5, 0.0, 0.0, 0.0]]]], device=device)
 
-    canted_surface_points = artist.util.utils.perform_canting(
+    canted_surface_points = transforms.perform_canting(
         canting_angles=canting, data=surface_points, device=device
     )
 
@@ -163,7 +164,7 @@ def test_find_span(device: torch.device) -> None:
     evaluation_points = artist.nurbs.utils.create_nurbs_evaluation_grid(
         torch.tensor([4, 5], device=device), device=device
     )
-    evaluation_points = artist.util.utils.normalize_points(evaluation_points)
+    evaluation_points = coordinates.normalize_points(evaluation_points)
 
     x, y = torch.meshgrid(
         torch.linspace(1e-2, 1 - 1e-2, 6),
