+++ tests/util/test_utils.py
@@ -11,527 +11,6 @@ from artist.util import utils
 
 
 @pytest.mark.parametrize(
-    "east_translation, north_translation, up_translation, expected",
-    [
-        (
-            torch.tensor([1.0]),
-            torch.tensor([2.0]),
-            torch.tensor([3.0]),
-            torch.tensor(
-                [
-                    [
-                        [1.0, 0.0, 0.0, 1.0],
-                        [0.0, 1.0, 0.0, 2.0],
-                        [0.0, 0.0, 1.0, 3.0],
-                        [0.0, 0.0, 0.0, 1.0],
-                    ]
-                ]
-            ),
-        ),
-        (
-            torch.tensor([1.0, 0.0, 0.0]),
-            torch.tensor([0.0, 1.0, 0.0, 1.0]),
-            torch.tensor([0.0, 0.0, 1.0, 1.0]),
-            None,
-        ),
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([0.0, 1.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0, 1.0]),
-            None,
-        ),
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([0.0, 1.0, 0.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0]),
-            None,
-        ),
-    ],
-)
-def test_translate_enu(
-    east_translation: torch.Tensor,
-    north_translation: torch.Tensor,
-    up_translation: torch.Tensor,
-    expected: torch.Tensor,
-    device: torch.device,
-) -> None:
-    """
-    Test that the correct translation matrix is created.
-
-    Parameters
-    ----------
-    east_translation : torch.Tensor
-        The translation in east direction.
-    north_translation : torch.Tensor
-        The translation in north direction.
-    up_translation : torch.Tensor
-        The translation in up direction.
-    expected : torch.Tensor
-        The expected overall translation or ``None`` if an error is expected.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    # Check if the ValueError is thrown as expected.
-    if expected is None:
-        with pytest.raises(ValueError) as exc_info:
-            utils.translate_enu(
-                e=east_translation.to(device),
-                n=north_translation.to(device),
-                u=up_translation.to(device),
-                device=device,
-            )
-        assert (
-            "The three tensors containing the east, north, and up translations must have the same shape."
-            in str(exc_info.value)
-        )
-    else:
-        # Check if the translation matrix is correct.
-        translation_matrix = utils.translate_enu(
-            e=east_translation.to(device),
-            n=north_translation.to(device),
-            u=up_translation.to(device),
-            device=device,
-        )
-        torch.testing.assert_close(
-            translation_matrix, expected.to(device), rtol=1e-4, atol=1e-4
-        )
-
-
-@pytest.mark.parametrize(
-    "point, expected",
-    [
-        (
-            torch.tensor([1.0, 0.0, 0.0]),
-            torch.tensor([1.0, 0.0, 0.0, 1.0]),
-        ),
-        (
-            torch.tensor([1.0, 0.0]),
-            None,
-        ),
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            None,
-        ),
-    ],
-)
-def test_3d_point_converter(
-    point: torch.Tensor, expected: torch.Tensor | None, device: torch.device
-) -> None:
-    """
-    Test the 3d to 4d point converter.
-
-    Parameters
-    ----------
-    point : torch.Tensor
-        A 3d point.
-    expected : torch.Tensor | None
-        A 4d point or ``None`` if an error is expected.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    # Check if the ValueError is thrown as expected.
-    if expected is None:
-        with pytest.raises(ValueError) as exc_info:
-            utils.convert_3d_points_to_4d_format(
-                points=point.to(device),
-                device=device,
-            )
-        assert f"Expected 3D points but got points of shape {point.shape}!" in str(
-            exc_info.value
-        )
-    else:
-        # Check if the 4d point is correct.
-        point_4d = utils.convert_3d_points_to_4d_format(
-            points=point.to(device),
-            device=device,
-        )
-        torch.testing.assert_close(point_4d, expected.to(device), rtol=1e-4, atol=1e-4)
-
-
-@pytest.mark.parametrize(
-    "direction, expected",
-    [
-        (
-            torch.tensor([1.0, 0.0, 0.0]),
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-        ),
-        (
-            torch.tensor([1.0, 0.0]),
-            None,
-        ),
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            None,
-        ),
-    ],
-)
-def test_3d_direction_converter(
-    direction: torch.Tensor, expected: torch.Tensor | None, device: torch.device
-) -> None:
-    """
-    Test the 3d to 4d point converter.
-
-    Parameters
-    ----------
-    direction : torch.Tensor
-        A 3d direction.
-    expected : torch.Tensor | None
-        A 4d direction or ``None`` if an error is expected.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    # Check if the ValueError is thrown as expected.
-    if expected is None:
-        with pytest.raises(ValueError) as exc_info:
-            utils.convert_3d_directions_to_4d_format(
-                directions=direction.to(device),
-                device=device,
-            )
-        assert (
-            f"Expected 3D directions but got directions of shape {direction.shape}!"
-            in str(exc_info.value)
-        )
-    else:
-        # Check if the 4d point is correct.
-        direction_4d = utils.convert_3d_directions_to_4d_format(
-            directions=direction.to(device),
-            device=device,
-        )
-        torch.testing.assert_close(
-            direction_4d, expected.to(device), rtol=1e-4, atol=1e-4
-        )
-
-
-@pytest.mark.parametrize(
-    "e_distortions, u_distortions, rays_to_rotate, expected_distorted_rays",
-    [
-        (  # Rotate single ray in the east direction.
-            torch.tensor([[[math.pi / 2]]]),
-            torch.tensor([[[0.0]]]),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor([[[[1.0, -1.0, 1.0, 0.0]]]]),
-        ),
-        (  # Rotate single ray in the up direction.
-            torch.tensor([[[0.0]]]),
-            torch.tensor([[[math.pi / 2]]]),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor([[[[-1.0, 1.0, 1.0, 0.0]]]]),
-        ),
-        (  # Rotate single ray in the east and then up direction.
-            torch.tensor([[[math.pi / 2]]]),
-            torch.tensor([[[math.pi / 2]]]),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor([[[[-1.0, -1.0, 1.0, 0.0]]]]),
-        ),
-        (  # Consider multiple rotations for a single ray in the east direction.
-            torch.tensor(
-                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
-            ),
-            torch.tensor([[[0]], [[0]], [[0]], [[0]]]),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor(
-                [
-                    [[[1.000000, -1.000000, 1.000000, 0.000000]]],
-                    [[[1.000000, -1.000000, -1.000000, 0.000000]]],
-                    [[[1.000000, 0.000000, 1.414214, 0.000000]]],
-                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for a single ray in the up direction.
-            torch.tensor([[[0]], [[0]], [[0]], [[0]]]),
-            torch.tensor(
-                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
-            ),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor(
-                [
-                    [[[-1.000000, 1.000000, 1.000000, 0.000000]]],
-                    [[[-1.000000, -1.000000, 1.000000, 0.000000]]],
-                    [[[0.000000, 1.414214, 1.000000, 0.000000]]],
-                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for a single ray in the east and then up direction.
-            torch.tensor(
-                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
-            ),
-            torch.tensor(
-                [[[math.pi / 2]], [[math.pi]], [[math.pi / 4]], [[2 * math.pi]]]
-            ),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            torch.tensor(
-                [
-                    [[[-1.000000, -1.000000, 1.000000, 0.000000]]],
-                    [[[-1.000000, 1.000000, -1.000000, 0.000000]]],
-                    [[[0.000000, 0.292893, 1.707107, 0.000000]]],
-                    [[[1.000000, 1.000000, 1.000000, 0.000000]]],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for multiple rays in the east direction.
-            torch.tensor(
-                [
-                    [[math.pi, math.pi / 2, math.pi / 4]],
-                    [[math.pi / 4, math.pi, math.pi / 2]],
-                ]
-            ),
-            torch.tensor([[[0, 0, 0]], [[0, 0, 0]]]),
-            torch.tensor(
-                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
-            ),
-            torch.tensor(
-                [
-                    [
-                        [
-                            [1.000000e00, -9.999999e-01, -1.000000e00, 0.000000e00],
-                            [2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [3.000000e00, -5.960464e-08, 4.242640e00, 0.000000e00],
-                        ]
-                    ],
-                    [
-                        [
-                            [1.000000e00, 0.000000e00, 1.414214e00, 0.000000e00],
-                            [2.000000e00, -2.000000e00, -2.000000e00, 0.000000e00],
-                            [3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ]
-                    ],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for multiple rays in the up direction.
-            torch.tensor([[[0, 0, 0]], [[0, 0, 0]]]),
-            torch.tensor(
-                [
-                    [[math.pi, math.pi / 2, math.pi / 4]],
-                    [[math.pi / 4, math.pi, math.pi / 2]],
-                ]
-            ),
-            torch.tensor(
-                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
-            ),
-            torch.tensor(
-                [
-                    [
-                        [
-                            [-9.999999e-01, -1.000000e00, 1.000000e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 4.242640e00, 3.000000e00, 0.000000e00],
-                        ]
-                    ],
-                    [
-                        [
-                            [0.000000e00, 1.414214e00, 1.000000e00, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-3.000000e00, 3.000000e00, 3.000000e00, 0.000000e00],
-                        ]
-                    ],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for multiple rays in the east and then up direction.
-            torch.tensor(
-                [
-                    [[math.pi, math.pi / 2, math.pi / 4]],
-                    [[math.pi / 4, math.pi, math.pi / 2]],
-                ]
-            ),
-            torch.tensor(
-                [
-                    [[math.pi, math.pi / 2, math.pi / 4]],
-                    [[math.pi / 4, math.pi, math.pi / 2]],
-                ]
-            ),
-            torch.tensor(
-                [[[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]]]
-            ),
-            torch.tensor(
-                [
-                    [
-                        [
-                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
-                        ]
-                    ],
-                    [
-                        [
-                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
-                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ]
-                    ],
-                ]
-            ),
-        ),
-        (  # Consider multiple rotations for multiple rays on four facets in the east and then up direction.
-            torch.tensor(
-                [
-                    [
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                    ],
-                    [
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                    ],
-                ]
-            ),
-            torch.tensor(
-                [
-                    [
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                        [math.pi, math.pi / 2, math.pi / 4],
-                    ],
-                    [
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                        [math.pi / 4, math.pi, math.pi / 2],
-                    ],
-                ]
-            ),
-            torch.tensor(
-                [
-                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
-                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
-                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
-                    [[1.0, 1.0, 1.0, 0.0], [2.0, 2.0, 2.0, 0.0], [3.0, 3.0, 3.0, 0.0]],
-                ]
-            ),
-            torch.tensor(
-                [
-                    [
-                        [
-                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
-                        ],
-                        [
-                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
-                        ],
-                        [
-                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
-                        ],
-                        [
-                            [-9.999999e-01, 1.000000e00, -9.999999e-01, 0.000000e00],
-                            [-2.000000e00, -2.000000e00, 2.000000e00, 0.000000e00],
-                            [-5.960464e-08, 8.786795e-01, 5.121320e00, 0.000000e00],
-                        ],
-                    ],
-                    [
-                        [
-                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
-                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ],
-                        [
-                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
-                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ],
-                        [
-                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
-                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ],
-                        [
-                            [0.000000e00, 2.928932e-01, 1.707107e00, 0.000000e00],
-                            [-2.000000e00, 2.000000e00, -2.000000e00, 0.000000e00],
-                            [-3.000000e00, -3.000000e00, 3.000000e00, 0.000000e00],
-                        ],
-                    ],
-                ]
-            ),
-        ),
-        (  # Test raise ValueError
-            torch.tensor([[[math.pi / 2]]]),
-            torch.tensor([[[0.0], [0.0]]]),
-            torch.tensor([[[1.0, 1.0, 1.0, 0.0]]]),
-            None,
-        ),
-    ],
-)
-def test_distortion_rotations(
-    e_distortions: torch.Tensor,
-    u_distortions: torch.Tensor,
-    rays_to_rotate: torch.Tensor,
-    expected_distorted_rays: torch.Tensor,
-    device: torch.device,
-) -> None:
-    """
-    Test the rotation function used for scattering rays by considering various rotations.
-
-    Parameters
-    ----------
-    e_distortions : torch.Tensor
-        The distortions in the east direction used in the rotation matrix.
-    u_distortions : torch.Tensor
-        The distortions in the upper direction used in the rotation matrix.
-    rays_to_rotate : torch.Tensor
-        The rays to rotate given the distortions.
-    expected_distorted_rays : torch.Tensor
-        The expected distorted rays after rotation.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    if expected_distorted_rays is None:
-        with pytest.raises(ValueError) as exc_info:
-            distorted_rays = (
-                utils.rotate_distortions(
-                    e=e_distortions.to(device),
-                    u=u_distortions.to(device),
-                    device=device,
-                )
-                @ rays_to_rotate.to(device).unsqueeze(-1)
-            ).squeeze(-1)
-        assert (
-            "The two tensors containing angles for the east and up rotation must have the same shape."
-            in str(exc_info.value)
-        )
-    else:
-        distorted_rays = (
-            utils.rotate_distortions(
-                e=e_distortions.to(device), u=u_distortions.to(device), device=device
-            )
-            @ rays_to_rotate.to(device).unsqueeze(-1)
-        ).squeeze(-1)
-
-        torch.testing.assert_close(distorted_rays, expected_distorted_rays.to(device))
-
-
-@pytest.mark.parametrize(
     "total_width, slope_width, plateau_width, expected",
     [
         (8, 2, 4, torch.tensor([0.25, 0.75, 1.0, 1.0, 1.0, 1.0, 0.75, 0.25])),
@@ -894,197 +373,4 @@ def test_azimuth_elevation_to_enu(
     )
     torch.testing.assert_close(
         enu_coordinates, expected.to(device), rtol=1e-4, atol=1e-4
-    )
-
-
-@pytest.mark.parametrize(
-    "from_orientation, to_orientation, expected_axis, expected_angle",
-    [
-        # Same orientation, no rotation, zero-degree angle.
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([1.0, 0.0, 0.0]),
-            torch.tensor(0.0),
-        ),
-        # From x-axis to y-axis, 90 degrees rotation around z.
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([0.0, 1.0, 0.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0]),
-            torch.tensor(torch.pi / 2),
-        ),
-        # From y-axis to z-axis, 90 degrees rotation around x.
-        (
-            torch.tensor([0.0, 1.0, 0.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0, 0.0]),
-            torch.tensor([1.0, 0.0, 0.0]),
-            torch.tensor(torch.pi / 2),
-        ),
-        # From positive x-axis, to negative x-axis, 180 degrees rotation, .
-        (
-            torch.tensor([1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([-1.0, 0.0, 0.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0]),
-            torch.tensor(torch.pi),
-        ),
-        # Non-normalized input vectors, from x-axis to y-axis, 90 degrees.
-        (
-            torch.tensor([2.0, 0.0, 0.0, 0.0]),
-            torch.tensor([0.0, 3.0, 0.0, 0.0]),
-            torch.tensor([0.0, 0.0, 1.0]),
-            torch.tensor(torch.pi / 2),
-        ),
-    ],
-)
-def test_rotation_angle_and_axis(
-    from_orientation: torch.Tensor,
-    to_orientation: torch.Tensor,
-    expected_axis: torch.Tensor,
-    expected_angle: torch.Tensor,
-    device: torch.device,
-) -> None:
-    """
-    Test the get rotation axis and angle method.
-
-    Parameters
-    ----------
-    from_orientation : torch.Tensor
-        The initial orientation (e.g., a 3D vector or rotation representation).
-    to_orientation : torch.Tensor
-        The target orientation to rotate into.
-    expected_axis : torch.Tensor
-        The expected unit vector representing the rotation axis.
-    expected_angle : torch.Tensor
-        The expected rotation angle in radians.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    axis, angle = utils.rotation_angle_and_axis(
-        from_orientation=from_orientation.to(device),
-        to_orientation=to_orientation.to(device),
-        device=device,
-    )
-
-    assert torch.allclose(axis, expected_axis.to(device), atol=1e-5)
-    assert torch.isclose(angle, expected_angle.to(device), atol=1e-5)
-
-
-@pytest.mark.parametrize(
-    "bitmap_coordinates, bitmap_resolution, target_area_indices, expected_coordinates",
-    [
-        (
-            torch.tensor([[127.5, 127.5], [63.75, 255.0], [0.0, 0.0]]),
-            torch.tensor([256, 256]),
-            torch.tensor([0, 0, 1]),
-            torch.tensor(
-                [
-                    [0.0000, 0.0000, 0.0000, 1.0000],
-                    [1.4941, 0.0000, -2.9883, 1.0000],
-                    [1.9961, 0.0000, 3.9922, 1.0000],
-                ]
-            ),
-        ),
-        (
-            torch.tensor([[127.5, 127.5], [127.5, 255.0], [0.0, 63.75]]),
-            torch.tensor([256, 256]),
-            torch.tensor([2, 2, 2]),
-            torch.tensor(
-                [
-                    [0.0000, 2.0000, 0.0000, 1.0000],
-                    [0.0000, 2.0000, -2.9883, 1.0000],
-                    [2.0000, 0.0123, 1.4941, 1.0000],
-                ]
-            ),
-        ),
-        (
-            torch.tensor([[255.0, 191.25], [255.0, 255.0]]),
-            torch.tensor([256, 256]),
-            torch.tensor([2, 0]),
-            torch.tensor(
-                [[-2.0000, 0.0123, -1.4941, 1.0000], [-2.9883, 0.0000, -2.9883, 1.0000]]
-            ),
-        ),
-    ],
-)
-def test_bitmap_coordinates_to_target_coordinates(
-    bitmap_coordinates: torch.Tensor,
-    bitmap_resolution: torch.Tensor,
-    target_area_indices: torch.Tensor,
-    expected_coordinates: torch.Tensor,
-    device: torch.device,
-) -> None:
-    """
-    Test the conversion from bitmap coordinates to target coordinates.
-
-    Parameters
-    ----------
-    bitmap_coordinates : torch.Tensor
-        The 2D pixel coordinates in the bitmap/image space.
-    bitmap_resolution : torch.Tensor
-        The resolution of the bitmap (e.g., width and height).
-    target_area_indices : torch.Tensor
-        Indices indicating which target area each bitmap coordinate maps to.
-    expected_coordinates : torch.Tensor
-        The expected 3D coordinates corresponding to the input bitmap coordinates.
-    device : torch.device
-        The device on which to initialize tensors.
-
-    Raises
-    ------
-    AssertionError
-        If test does not complete as expected.
-    """
-    mock_solar_tower = MagicMock(spec=SolarTower)
-    mock_target_areas_planar = MagicMock(spec=TowerTargetAreasPlanar)
-    mock_target_areas_planar.names = ["planar1", "planar2"]
-    mock_target_areas_planar.dimensions = torch.tensor(
-        [[6.0, 6.0], [2.0, 4.0]], device=device
-    )
-    mock_target_areas_planar.centers = torch.tensor(
-        [[0.0, 0.0, 0.0, 1.0], [1.0, 0.0, 2.0, 1.0]], device=device
-    )
-    mock_target_areas_cylindrical = MagicMock(spec=TowerTargetAreasCylindrical)
-    mock_target_areas_cylindrical.names = ["cylinder1"]
-    mock_target_areas_cylindrical.normals = torch.tensor(
-        ([[0.0, 1.0, 0.0, 0.0]]), device=device
-    )
-    mock_target_areas_cylindrical.axes = torch.tensor(
-        ([[0.0, 0.0, 1.0, 0.0]]), device=device
-    )
-    mock_target_areas_cylindrical.radii = torch.tensor(([[2.0]]), device=device)
-    mock_target_areas_cylindrical.heights = torch.tensor(([[6.0]]), device=device)
-    mock_target_areas_cylindrical.opening_angles = torch.tensor(
-        ([[math.pi]]), device=device
-    )
-    mock_target_areas_cylindrical.centers = torch.tensor(
-        ([[0.0, 0.0, 0.0, 1.0]]), device=device
-    )
-
-    mock_solar_tower.target_areas = [
-        mock_target_areas_planar,
-        mock_target_areas_cylindrical,
-    ]
-    mock_solar_tower.number_of_target_area_types = 2
-    mock_solar_tower.number_of_target_areas_per_type = torch.tensor(
-        [2, 1], device=device
-    )
-    mock_solar_tower.target_name_to_index = {"planar1": 0, "planar2": 1, "cylinder1": 2}
-    mock_solar_tower.index_to_target_area = {0: "planar1", 1: "planar2", 2: "cylinder1"}
-
-    target_coordinates = utils.bitmap_coordinates_to_target_coordinates(
-        bitmap_coordinates=bitmap_coordinates.to(device),
-        bitmap_resolution=bitmap_resolution.to(device),
-        solar_tower=mock_solar_tower,
-        target_area_indices=target_area_indices.to(device),
-        device=device,
-    )
-
-    torch.testing.assert_close(
-        target_coordinates, expected_coordinates.to(device), rtol=1e-4, atol=1e-4
     )
